moved files to pre-conversion-subs
