+++ packages/unified-latex-to-pretext/libs/pre-conversion-subs/expand-user-defined-macros.ts
