+++ packages/unified-latex-to-pretext/tests/report-unsupported-macro-katex.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 import util from "util";
 import { getParser } from "@unified-latex/unified-latex-util-parse";
-import { reportMacrosUnsupportedByKatex } from "@unified-latex/unified-latex-to-pretext/libs/report-unsupported-macro-katex";
+import { reportMacrosUnsupportedByKatex } from "@unified-latex/unified-latex-to-pretext/libs/pre-conversion-subs/report-unsupported-macro-katex";
 
 // Make console.log pretty-print by default
 const origLog = console.log;
@@ -63,6 +63,9 @@ describe("unified-latex-to-pretext:report-unsupported-macro-katex", () => {
         const parser = getParser();
         const ast = parser.parse(value);
 
-        expect(reportMacrosUnsupportedByKatex(ast)).toEqual(["unsupported", "baz"]);
+        expect(reportMacrosUnsupportedByKatex(ast)).toEqual([
+            "unsupported",
+            "baz",
+        ]);
     });
 });
